refactor: split allowlist module to prevent circular imports
Split the allowlist module into two separate modules to break the
circular import chain that caused ImportError on Python 3.14:

allowlist.py → helpers/testing_farm.py → checker/testing_farm.py
→ handlers/__init__.py → handlers/forges.py → allowlist.py

Changes:

Core module (allowlist.py):
- Removed job helper imports (CoprBuildJobHelper, TestingFarmJobHelper)
- Kept all core allowlist management methods:
  - Static methods for checking/managing namespaces
  - FAS authentication methods
  - Namespace approval/denial methods

New event checker module (allowlist_checker.py):
- Created AllowlistChecker class extending Allowlist
- Moved all event checking methods that depend on job helpers:
  - check_and_report() and all _check_*() event methods
  - _check_pr_report_status() which uses job helpers
- Imports job helpers only in this module

Updated imports:
- jobs.py: now uses AllowlistChecker for check_and_report()
- Test files: updated to use AllowlistChecker where needed
- handlers/forges.py: continues to use Allowlist (core methods only)

Benefits over local imports (PR #3178):
1. Cleaner separation of concerns: Core allowlist logic vs event checking
2. No runtime overhead: Local imports are checked on every call
3. Better maintainability: Clear dependency boundaries
4. Future-proof: Works with all Python versions, including 3.14+

The allowlist.py module can now be imported independently without
triggering the circular import chain.
