+++ tests/integration/test_commit_comment.py
@@ -20,7 +20,7 @@ from packit_service.models import (
     ProjectEventModel,
     ProjectEventModelType,
 )
-from packit_service.worker.allowlist import Allowlist
+from packit_service.worker.allowlist_checker import AllowlistChecker
 from packit_service.worker.helpers.build import CoprBuildJobHelper
 from packit_service.worker.helpers.testing_farm import TestingFarmJobHelper
 from packit_service.worker.jobs import SteveJobs
@@ -82,7 +82,7 @@ def mock_commit_comment_functionality(request):
     ).and_return(db_project_object)
     flexmock(LocalProject, refresh_the_arguments=lambda: None)
     flexmock(LocalProjectBuilder, _refresh_the_state=lambda *args: flexmock())
-    flexmock(Allowlist, check_and_report=True)
+    flexmock(AllowlistChecker, check_and_report=True)
 
 
 @pytest.mark.parametrize(
